requirements.txt
streamlit
openai
pandas
scikit-learn
pymupdf
spacy
python-dotenv
